+++ ppci/graph/algorithm/__init__.py
@@ -1,3 +1,2 @@
 
 """ Package containing several graph algorithms """
-