+++ ppci/lang/tools/regex/symbol_set.py
@@ -6,4 +6,3 @@ class SymbolSet:
 
     def __contains__(self, item):
         return item in self._symbols
-