+++ tools/ppci_explorer.py
@@ -16,43 +16,34 @@ Idea:
 
 import io
 import logging
+from itertools import cycle
 
 from prompt_toolkit import Application
 import prompt_toolkit as pt
 from pygments.styles import get_style_by_name
-from prompt_toolkit.styles import style_from_pygments_cls
-from pygments.token import Token
-from prompt_toolkit.key_binding import KeyBindings
 from pygments.lexers import CLexer
-from prompt_toolkit.lexers import PygmentsLexer
+
 from prompt_toolkit.buffer import Buffer
+from prompt_toolkit.key_binding import KeyBindings
+from prompt_toolkit.key_binding.bindings.focus import focus_next
+from prompt_toolkit.key_binding.bindings.focus import focus_previous
 from prompt_toolkit.keys import Keys
+from prompt_toolkit.lexers import PygmentsLexer
 from prompt_toolkit.layout.containers import HSplit, Window, VSplit
 from prompt_toolkit.layout.containers import ConditionalContainer
-from prompt_toolkit.layout.dimension import LayoutDimension as D
 from prompt_toolkit.layout.controls import BufferControl
 from prompt_toolkit.layout.controls import FormattedTextControl
 from prompt_toolkit.layout import ScrollbarMargin, NumberedMargin
-from prompt_toolkit.widgets import TextArea, Frame
+from prompt_toolkit.widgets import Frame
 from prompt_toolkit.layout.layout import Layout
 from prompt_toolkit.layout.processors import Processor, Transformation
 from prompt_toolkit.filters import Condition
+from prompt_toolkit.styles import style_from_pygments_cls
 
 from ppci import __version__ as ppci_version
 from ppci import api
 from ppci.binutils.outstream import TextOutputStream
 from ppci.common import CompilerError, logformat
-
-
-def get_title_bar_tokens():
-    return \
-         'Welcome to the ppci explorer version {} (prompt_toolkit {})'.format(ppci_version, pt.__version__) + \
-         ' [Arch = {} (F7)] '.format(compiler.arch) + \
-         ' [Optimize = {} (F8)] '.format(compiler.optimize)
-
-
-def get_help_tokens():
-    return 'F9=toggle log F10=exit '
 
 
 class MyHandler(logging.Handler):
@@ -69,48 +60,122 @@ class DisplayErrorsProcessor(Processor):
     def __init__(self):
         self.errors = {}
 
-    def apply_transformation(self, cli, document, lineno, source_to_display, tokens):
-        tokens = list(tokens)
-        if lineno + 1 in self.errors:
-            tokens.append((Token.Title,
-                           '// {}'.format(self.errors[lineno + 1])))
+    def apply_transformation(self, transformation_input):
+        tokens = list(transformation_input.fragments)
+        lineno = transformation_input.lineno + 1
+        if lineno in self.errors:
+            tokens.append(('',
+                           '// {}'.format(self.errors[lineno])))
         return Transformation(tokens)
 
 
-def arch_gen(archs):
-    """ Endless sequence of architectures """
-    while True:
-        for a in archs:
-            yield a
-
-
-class CodeEdit:
+class PpciExplorer:
+    """ Ppci explorer. """
     def __init__(self):
-        self.buffer = Buffer()
+        available_archs = ['arm', 'x86_64', 'riscv', 'avr', 'or1k', 'xtensa']
+        # State variables:
+        self.arch = available_archs[-1]
+        self.show_log = False
+        self.optimize = False
+
+        self.archs = cycle(available_archs)
+
+        # Some key bindings:
+        kb = KeyBindings()
+
+        @kb.add(Keys.F10, eager=True)
+        def quit_(event):
+            event.app.exit()
+
+        kb.add(Keys.F7, eager=True)(self.next_architecture)
+        kb.add(Keys.F8, eager=True)(self.toggle_optimize)
+        kb.add(Keys.F9, eager=True)(self.toggle_log)
+        kb.add('tab')(focus_next)
+        kb.add('s-tab')(focus_previous)
+
+        log_buffer = Buffer(multiline=True)
+
+        show_log = Condition(lambda: self.show_log)
+        self.errors_processor = DisplayErrorsProcessor()
+
+        # Source:
+        self.source_buffer = Buffer(multiline=True)
         src_lexer = PygmentsLexer(CLexer)
-        self.control = BufferControl(
-            buffer=self.buffer,
-            lexer=src_lexer,
-        )
-        self.window = Window(
-            content=self.control,
+        code_edit = Window(
+            content=BufferControl(
+                buffer=self.source_buffer, lexer=src_lexer,
+                input_processors=[self.errors_processor],
+            ),
             left_margins=[NumberedMargin()],
             right_margins=[ScrollbarMargin(display_arrows=True)],
         )
 
-    def __pt_container__(self):
-        return self.window
+        # Output:
+        self.output_buffer = Buffer(multiline=True)
+        result_window = Window(
+            content=BufferControl(self.output_buffer),
+            right_margins=[ScrollbarMargin(display_arrows=True)],
+        )
+        layout = Layout(HSplit([
+            Window(FormattedTextControl(self.get_title_bar_tokens), height=1),
+            VSplit([
+                Frame(body=code_edit, title="source code"),
+                Frame(body=result_window, title='assembly output'),
+                ConditionalContainer(
+                    Window(content=BufferControl(log_buffer)),
+                    filter=show_log
+                ),
+            ]),
+            Window(
+                FormattedTextControl(text='F9=toggle log F10=exit'),
+                height=1, style='class:status'
+            ),
+        ]))
 
+        style = style_from_pygments_cls(get_style_by_name('vim'))
+        log_handler = MyHandler(log_buffer)
+        fmt = logging.Formatter(fmt=logformat)
+        log_handler.setFormatter(fmt)
+        log_handler.setLevel(logging.DEBUG)
+        logging.getLogger().setLevel(logging.INFO)
+        logging.getLogger().addHandler(log_handler)
 
-class Compiler:
-    def __init__(self):
-        self.archs = arch_gen([
-            'arm', 'x86_64', 'riscv', 'avr', 'or1k', 'xtensa'])
-        self.next_architecture()
-        self.show_log = False
-        self.optimize = False
+        self.source_buffer.on_text_changed += self.on_change
+
+        self.application = Application(
+            layout=layout,
+            key_bindings=kb,
+            style=style,
+            full_screen=True)
+
+        self.source_buffer.text = DEMO_SOURCE
+
+    def on_change(self, source_buffer):
+        self.do_compile()
+
+    def get_title_bar_tokens(self):
+        return \
+             'Welcome to the ppci explorer {}'.format(ppci_version) + \
+             '(prompt_toolkit {})'.format(pt.__version__) + \
+             ' [Arch = {} (F7)] '.format(self.arch) + \
+             ' [Optimize = {} (F8)] '.format(self.optimize)
+
+    def do_compile(self):
+        """ Try a compilation. """
+        self.errors_processor.errors.clear()
+        try:
+            self.output_buffer.text = self.compile(self.source_buffer.text)
+        except CompilerError as ex:
+            if ex.loc:
+                self.errors_processor.errors[ex.loc.row] = ex.msg
+                self.output_buffer.text = ''
+            else:
+                self.output_buffer.text = str(ex)
+        except Exception as ex:  # Catch the more hard-core exceptions.
+            self.output_buffer.text = str(ex)
 
     def compile(self, source):
+        """ Compile the given source with current settings. """
         f = io.StringIO(source)
         ir_module = api.c_to_ir(f, self.arch)
         f2 = io.StringIO()
@@ -120,104 +185,31 @@ class Compiler:
         api.ir_to_stream(ir_module, self.arch, text_stream)
         return f2.getvalue()
 
-    def next_architecture(self):
+    def next_architecture(self, event):
         self.arch = next(self.archs)
+        self.do_compile()
 
-    def toggle_log(self):
-        self.show_log = not self.show_log
-
-    def toggle_optimize(self):
+    def toggle_optimize(self, event):
         self.optimize = not self.optimize
+        self.do_compile()
 
-
-compiler = Compiler()
+    def toggle_log(self, event):
+        self.show_log = not self.show_log
 
 
 def ppci_explorer():
-    kb = KeyBindings()
+    """ Run the ppci explorer. """
+    explorer = PpciExplorer()
+    explorer.application.run()
 
-    @kb.add(Keys.F10, eager=True)
-    def quit_(event):
-        event.app.exit()
 
-    @kb.add(Keys.F7, eager=True)
-    def next_arch_(event):
-        compiler.next_architecture()
-        do_compile()
+DEMO_SOURCE = """
+int g=23;
 
-    @kb.add(Keys.F9, eager=True)
-    def toggle_log_(event):
-        compiler.toggle_log()
-
-    @kb.add(Keys.F8, eager=True)
-    def toggle_optimize_(event):
-        compiler.toggle_optimize()
-        do_compile()
-
-    buffers = {
-        'output': Buffer(multiline=True),
-        'logs': Buffer(multiline=True),
-    }
-
-    show_log = Condition(lambda: compiler.show_log)
-    errors_processor = DisplayErrorsProcessor()
-    code_edit = CodeEdit()
-    source_buffer = code_edit.buffer
-    layout = Layout(HSplit([
-        Window(FormattedTextControl(get_title_bar_tokens), height=1),
-        VSplit([
-            Frame(
-                body=code_edit, title="source code",
-            ),
-            Frame(
-                body=Window(content=BufferControl(buffers['output'])),
-                title='assembly output'
-            ),
-            ConditionalContainer(
-                Window(content=BufferControl(buffers['logs'])),
-                filter=show_log),
-        ]),
-        Window(FormattedTextControl(get_help_tokens), height=1),
-    ]))
-
-    style = style_from_pygments_cls(get_style_by_name('vim'))
-    log_handler = MyHandler(buffers['logs'])
-    fmt = logging.Formatter(fmt=logformat)
-    log_handler.setFormatter(fmt)
-    log_handler.setLevel(logging.DEBUG)
-    logging.getLogger().setLevel(logging.INFO)
-    logging.getLogger().addHandler(log_handler)
-
-    def on_change(source_buffer):
-        do_compile()
-
-    def do_compile():
-        errors_processor.errors.clear()
-        try:
-            buffers['output'].text = compiler.compile(source_buffer.text)
-        except CompilerError as ex:
-            if ex.loc:
-                errors_processor.errors[ex.loc.row] = ex.msg
-                buffers['output'].text = ''
-            else:
-                buffers['output'].text = str(ex)
-
-    source_buffer.on_text_changed += on_change
-
-    application = Application(
-        layout=layout,
-        key_bindings=kb,
-        full_screen=True)
-
-    source_buffer.text = """
-    int g=23;
-
-    int add(int a, int b) {
-      return a + b - g;
-    }
-    """
-
-    application.run()
+int add(int a, int b) {
+  return a + b - g;
+}
+"""
 
 
 if __name__ == '__main__':