+++ docs/changelog.rst
@@ -12,6 +12,10 @@ Release 0.5.7 (Upcoming)
 
 * Added support for the m68k processor.
 * Added support for the microblaze processor.
+* Added RISCV floating point support (by Michael).
+* Add S-record format.
+* Add amiga hunk format.
+* Add OCaml bytecode file reader.
 
 Release 0.5.6 (Aug 22, 2018)
 ----------------------------